refactor: pull logout modal to its own file, make compliant w/ design
